Move resource models to block-kit
