+++ features/lessons/src/main/kotlin/app/ss/lessons/components/spec/CreditSpec.kt
@@ -22,7 +22,7 @@
 
 package app.ss.lessons.components.spec
 
-import app.ss.models.Credit
+import io.adventech.blockkit.model.resource.Credit
 import java.util.UUID
 import javax.annotation.concurrent.Immutable
