+++ features/resource/src/main/kotlin/ss/resource/components/content/ResourceSectionsStateProducer.kt
@@ -28,14 +28,14 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.unit.dp
-import app.ss.models.feed.FeedResourceKind
-import app.ss.models.resource.Resource
-import app.ss.models.resource.ResourceDocument
-import app.ss.models.resource.ResourceSection
-import app.ss.models.resource.ResourceSectionViewType
 import com.slack.circuit.retained.rememberRetained
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
+import io.adventech.blockkit.model.feed.FeedResourceKind
+import io.adventech.blockkit.model.resource.Resource
+import io.adventech.blockkit.model.resource.ResourceDocument
+import io.adventech.blockkit.model.resource.ResourceSection
+import io.adventech.blockkit.model.resource.ResourceSectionViewType
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
 import org.joda.time.DateTime
