+++ features/resource/src/main/kotlin/ss/resource/components/spec/FeatureSpec.kt
@@ -22,7 +22,7 @@
 
 package ss.resource.components.spec
 
-import app.ss.models.Feature
+import io.adventech.blockkit.model.resource.Feature
 import javax.annotation.concurrent.Immutable
 
 @Immutable
