+++ libraries/block-kit/model/src/main/kotlin/io/adventech/blockkit/model/feed/FeedType.kt
@@ -20,36 +20,18 @@
  * THE SOFTWARE.
  */
 
-package app.ss.models.resource
+package io.adventech.blockkit.model.feed
 
 import androidx.annotation.Keep
-import app.ss.models.Credit
-import app.ss.models.Feature
-import app.ss.models.feed.FeedResourceKind
-import app.ss.models.feed.FeedType
+import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @Keep
-@JsonClass(generateAdapter = true)
-data class Resource(
-    val id: String,
-    val name: String,
-    val title: String,
-    val startDate: String?,
-    val endDate: String?,
-    val description: String?,
-    val introduction: String?,
-    val index: String,
-    val type: FeedType,
-    val credits: List<Credit>,
-    val features: List<Feature>,
-    val primaryColor: String,
-    val primaryColorDark: String,
-    val subtitle: String?,
-    val covers: ResourceCovers,
-    val kind: FeedResourceKind,
-    val sectionView: ResourceSectionViewType?,
-    val sections: List<ResourceSection>?,
-    val cta: ResourceCTA?,
-    val preferredCover: ResourcePreferredCover?,
-)
+@JsonClass(generateAdapter = false)
+enum class FeedType {
+    UNKNOWN,
+    @Json(name = "ss") SS,
+    @Json(name = "aij") AIJ,
+    @Json(name = "devo") DEVO,
+    @Json(name = "pm") PM,
+}
