+++ libraries/block-kit/model/src/main/kotlin/io/adventech/blockkit/model/resource/ResourceCTA.kt
@@ -20,29 +20,14 @@
  * THE SOFTWARE.
  */
 
-package app.ss.models.resource
+package io.adventech.blockkit.model.resource
 
 import androidx.annotation.Keep
-import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
-enum class ResourceCoverType(val aspectRatio: Float) {
-    LANDSCAPE(1280f / 720f),
-    PORTRAIT(854f / 1280f),
-    SPLASH(1f),
-    SQUARE(1f)
-}
-
 @Keep
-@JsonClass(generateAdapter = false)
-enum class ResourcePreferredCover {
-    UNKNOWN,
-    @Json(name = "landscape")
-    LANDSCAPE,
-    @Json(name = "portrait")
-    PORTRAIT,
-    @Json(name = "splash")
-    SPLASH,
-    @Json(name = "square")
-    SQUARE
-}
+@JsonClass(generateAdapter = true)
+data class ResourceCTA(
+    val hidden: Boolean?,
+    val text: String?
+)
